% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{methods}
\name{DB2-schema}
\alias{DB2-schema}
\alias{dbExistsSchema,DB2Connection,character-method}
\alias{dbListSchemas,DB2Connection-method}
\title{DB2 Schema-related functions}
\usage{
\S4method{dbExistsSchema}{DB2Connection,character}(conn, schema, ...)

\S4method{dbListSchemas}{DB2Connection}(conn, sys.include = FALSE)
}
\description{
DB2 Schema-related functions
}

