% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\docType{methods}
\name{DB2-tables}
\alias{DB2-tables}
\alias{dbExistsTable,DB2Connection,character-method}
\alias{dbFindTables,DB2Connection,character-method}
\alias{dbListFields,DB2Connection,character-method}
\alias{dbListTables,DB2Connection-method}
\alias{dbReadTable,DB2Connection,character-method}
\title{Convenience functions for reading/writing DB2 tables}
\usage{
\S4method{dbReadTable}{DB2Connection,character}(conn, name,
  check.names = TRUE, select.cols = "*", schema = "syscat")

\S4method{dbListTables}{DB2Connection}(conn, schemas = NULL, ...)

\S4method{dbExistsTable}{DB2Connection,character}(conn, name, schema = NULL)

\S4method{dbFindTables}{DB2Connection,character}(conn, name, schema = NULL,
  sys.include = FALSE)

\S4method{dbListFields}{DB2Connection,character}(conn, name,
  schema = "syscat")
}
\arguments{
\item{conn}{a \code{\linkS4class{DB2Connection}} object, produced by
\code{\link[DBI]{dbConnect}}}

\item{name}{a character string specifying a table name with schema, e.g syscat.tables. DB2 table names
are \emph{not} case sensitive, e.g., table names \code{ABC} and \code{abc} are considered equal.}

\item{check.names}{If \code{TRUE}, the default, column names will be
converted to valid R identifiers.}

\item{select.cols}{A SQL statement, either in the form of
1) a character vector of length 1, e.g. "x, y, z"
2) a character vector of length > 1, e.g. c("x", "y", "z")
3) an expression using ..(), e.g. ..(x, y, z)
"*" selects all columns,}

\item{schema}{a character. If name does not include schema (ie. not like syscat.tables),
user must provide schema through this parameter, or by default use "syscat"}

\item{conn}{a \code{\linkS4class{DB2Connection}} object, produced by \code{\link[DBI]{dbConnect}}}

\item{name, }{schema a character vector of length 1 (extra entries will be ignored)}

\item{conn}{An existing \code{\linkS4class{DB2Connection}}}

\item{name}{a length 1 character vector giving the name of a table.}
}
\value{
A data.frame in the case of \code{dbReadTable};
}
\description{
Convenience functions for reading/writing DB2 tables

Read the table

List fields in specified table.
}
\note{
Note that the data.frame returned by \code{dbReadTable} only has
primitive data, e.g., it does not coerce character data to factors.
}
\examples{
\dontrun{
con <- dbConnect(DB2(), user = user, password = password)
dbListFields(con, "TABNAME")
dbDisconnect(con)
}
}

