% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{DB2Connection-class}
\alias{DB2Connection-class}
\alias{dbConnect,DB2Driver-method}
\alias{dbDisconnect,DB2Connection-method}
\alias{dbGetInfo,DB2Connection-method}
\alias{dbIsValid,DB2Connection-method}
\alias{show,DB2Connection-method}
\title{DB2 connection class.}
\usage{
\S4method{dbConnect}{DB2Driver}(drv, dbname = NULL, host = NULL,
  port = 50000, user = NULL, password = NULL, ...)

\S4method{dbGetInfo}{DB2Connection}(dbObj, ...)

\S4method{dbIsValid}{DB2Connection}(dbObj, ...)

\S4method{show}{DB2Connection}(object)

\S4method{dbDisconnect}{DB2Connection}(conn)
}
\arguments{
\item{drv}{An object created by \code{DB2()}}
}
\description{
DB2 connection class.

DB2 Connection Functions
}
\examples{
\dontrun{
db <- dbConnect(DB2::DB2())
dbWriteTable(db, "mtcars", mtcars)
dbGetQuery(db, "SELECT * FROM mtcars WHERE cyl == 4")
}
}
\keyword{internal}

