% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{methods}
\name{DB2Driver-class}
\alias{DB2Driver-class}
\alias{dbGetInfo,DB2Driver-method}
\alias{dbIsValid,DB2Driver-method}
\alias{dbUnloadDriver,DB2Driver-method}
\alias{show,DB2Driver-method}
\title{DB2 database driver functions.}
\usage{
\S4method{dbUnloadDriver}{DB2Driver}(drv, ...)

\S4method{show}{DB2Driver}(object)

\S4method{dbGetInfo}{DB2Driver}(dbObj, ...)

\S4method{dbIsValid}{DB2Driver}(dbObj, ...)
}
\description{
DB2 database driver functions.
}

