% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\docType{class}
\name{DB2Result-class}
\alias{DB2Result}
\alias{DB2Result-class}
\alias{dbClearResult,DB2Result-method}
\alias{dbColumnInfo,DB2Result-method}
\alias{dbFetch,DB2Result-method}
\alias{dbGetInfo,DB2Result-method}
\alias{dbGetRowCount,DB2Result-method}
\alias{dbGetStatement,DB2Result-method}
\alias{dbHasCompleted,DB2Result-method}
\alias{dbIsValid,DB2Result-method}
\alias{dbSendQuery,DB2Connection,character-method}
\title{DB2 results class.}
\usage{
\S4method{dbSendQuery}{DB2Connection,character}(conn, statement, ...)

\S4method{dbClearResult}{DB2Result}(res, ...)

\S4method{dbFetch}{DB2Result}(res, n = -1, stringsAsFactors = FALSE, ...)

\S4method{dbHasCompleted}{DB2Result}(res, ...)

\S4method{dbGetInfo}{DB2Result}(dbObj, ...)

\S4method{dbIsValid}{DB2Result}(dbObj, ...)

\S4method{dbGetStatement}{DB2Result}(res, ...)

\S4method{dbColumnInfo}{DB2Result}(res, ...)

\S4method{dbGetRowCount}{DB2Result}(res, ...)
}
\description{
DB2 results class.

Send a query to DB2

Retrieve records from DB2 query
}
\examples{
# This is another good place to put examples
}
\keyword{internal}

