% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{quote}
\alias{dbQuoteIdentifier,DB2Connection,SQL-method}
\alias{dbQuoteIdentifier,DB2Connection,character-method}
\alias{dbQuoteString,DB2Connection,SQL-method}
\alias{dbQuoteString,DB2Connection,character-method}
\alias{quote}
\title{Quote DB2 strings and identifiers.}
\usage{
\S4method{dbQuoteString}{DB2Connection,character}(conn, x, ...)

\S4method{dbQuoteString}{DB2Connection,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{DB2Connection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{DB2Connection,SQL}(conn, x, ...)
}
\arguments{
\item{conn}{A \linkS4class{DB2Connection} created by \code{dbConnect()}}

\item{x}{A character to escaped}

\item{...}{Other arguments needed for compatibility with generic}
}
\description{
Quote DB2 strings and identifiers.
}
\examples{
library(DBI)
con <- dbConnect(RDB2::DB2())

x <- c("a", "b c", "d'e", "\\\\f")
dbQuoteString(con, x)
dbQuoteIdentifier(con, x)
}

