% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{trimws}
\alias{trimws}
\alias{trimws,character-method}
\title{Remove whitespaces of character strings from a data.frame}
\usage{
\S4method{trimws}{data.frame}(x, stringsAsFactors = FALSE, ...)

\S4method{trimws}{character}(x, which = c("both", "left", "right"), ...)
}
\arguments{
\item{x}{a data.frame with character vectors or a character}

\item{stringAsFactors}{a logical}
}
\description{
Remove whitespaces of character strings from a data.frame
}
\note{
Each column of the data.frame must be a character vectior. This function will convert it to character if it is not.
}

